refactor(chatbot): rilocate chatbot classes in package (#24)
* create package chatbot and rilocate classes Cadocs and IntentManager
* update import
* update tests

Closes #21
